requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.1.0
pymupdf>=1.23.0
pillow>=10.0.0
pdf2image>=1.16.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
img2table>=1.4.0
openpyxl>=3.0.0
pandas>=2.0.0
tabulate>=0.9.0
matplotlib>=3.5.0
reportlab>=3.6.0
streamlit>=1.32.0 